% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmfrailty.R
\name{emmfrailty}
\alias{emmfrailty}
\title{Fitting a semiparametric multivariate mixed-Poisson model}
\usage{
emmfrailty(formula, data, frailty = NULL, paircop = TRUE,
  margins = "lnorm", ntype, twostage = FALSE, model = FALSE,
  model.matrix = FALSE, control = list(), init = list(),
  eventnames = NULL, parallel = FALSE, ncore = NULL, ...)
}
\arguments{
\item{formula}{a formula object with an obect of the type \code{Surv} on the left side and the terms on the right side. Note that \code{+strata()} is not supported.}

\item{data}{a 'data.frame' which has variables in 'formula'}

\item{frailty}{a charhacter string specifying a group}

\item{paircop}{a logical value: if \code{TRUE}, a dependent pairwise likelihood is used; otherwise, an independent mixed-Poissom model is used.}

\item{ntype}{the number of types of events.}

\item{twostage}{a logical value: if \code{TRUE}, two-stage procedure is used.}

\item{model}{logical value: if \code{TRUE}, the model frame is returned.}

\item{model.matrix}{logical value: if \code{TRUE}, the x matirx is returned.}

\item{control}{an object of class specifying control options created by \code{\link{controlList}}.}

\item{init}{a list specifying inital values of a vector of variances of random effects and dependence parameters. Default initial value for a variance of random effect is 1 and 0 for the dependence parameter.}

\item{eventnames}{a vector of character string speficying event names}

\item{parallel}{logical value: if \code{TRUE}, it supports parallel execution.}

\item{ncore}{the number of cores if \code{parallel = TRUE}.}
}
\value{
an object of class \code{emmfrailty} representing the fit.
\item{beta.coefficients}{a list of each type of events of the estimated regressoin coefficients.}
\item{sig2}{a vector of the estimated variances of random effects.}
\item{rho}{a vector of the estimated dependence parameters from a Gaussian copula function.}
\item{ktau}{a vector of the estimated Kendall's tau from a Gaussian copula function.}
\item{basecmhaz}{a list of each type of events of the estiamted cumulative baseline hazards.}
\item{logLik}{a value of log likelihood at the final values of the parameters.}
\item{iter}{the number of iterations used.}
\item{conv}{an integer code for the convergence. 0 indicates successful convergence, 1 indicates a failure to convergence, and 2 indicates the estimated dependence parameter reaches the boundary. }
\item{var}{a variance-covariance matrix of the estimates.}
\item{betavar}{a list of each type of the variances of the regression coefficient estimates.}
\item{sig2var}{a vector of the variances of random effects variances estimates.}
\item{rhovar}{a vector of the dependence parameter estimates.}
\item{varnames}{a list of event names and variable names.}
\item{control}{a list of control arguments used.}
\item{n}{the number of sample size.}
\item{nevent}{a vector of the number of each types of events.}
\item{neventtype}{the number of types of events.}
\item{nknot}{the number nodes used in Gaussian-quadrature.}
\item{margins}{a vector of character strings specyting marginal distributions of random effects.}

The object will also contain the following: two_stage, copula, model, call, optionally x, and model.
}
\description{
Fitting a semiparametric multivariate mixed-Poisson model via a Gaussian copula with a log-normal marginal distribution of random effects
}
