% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parmfrailty.R
\name{parmfrailty}
\alias{parmfrailty}
\title{Fitting a parametric multivariate mixed-Poisson model}
\usage{
parmfrailty(formula, data, frailty = NULL, dist = "weibull",
  margins = "lnorm", ntype, model = FALSE, model.matrix = FALSE,
  control = list(), init = list(), eventnames = NULL,
  parallel = FALSE, ncore = NULL, ...)
}
\arguments{
\item{formula}{a formula object with an obect of the type \code{Surv} on the left side and the terms on the right side. Note that \code{+strata()} is not supported.}

\item{data}{a 'data.frame' which has variables in 'formula'}

\item{frailty}{a charhacter string specifying a group}

\item{dist}{a vector of character strings specifying a distribution for the baseline for each type of events. If it is a character string, it applies to all types of events. Default is "weibull". Other options are "ev", "loglogistic", and "lognormal". See \code{phreg}.}

\item{ntype}{the number of types of events.}

\item{model}{logical value: if \code{TRUE}, the model frame is returned.}

\item{model.matrix}{logical value: if \code{TRUE}, the x matirx is returned.}

\item{control}{an object of class specifying control options created by \code{\link{controlList}}.}

\item{init}{a list specifying inital values of a vector of variances of random effects and dependence parameters. Default initial value for a variance of random effect is 1 and 0 for the dependence parameter.}

\item{eventnames}{a vector of character string speficying event names}

\item{parallel}{logical value: if \code{TRUE}, it supports parallel execution.}

\item{ncore}{the number of cores if \code{parallel = TRUE}.}
}
\value{
an object of class \code{parmfrailty} representing the fit. 
\item{beta.coefficients}{a list of each types of events of the estimated regressoin coefficients.}
\item{logalpha.coefficients}{a list of each types of events of the estimated log of baseline parameters.}
\item{sig2}{a vector of the estimated variances of random effects.}
\item{rho}{a vector of the estimated dependence parameters from a Gaussian copula function.}
\item{ktau}{a vector of the estimated Kendall's tau from a Gaussian copula function.}
\item{logLik}{a value of log likelihood at the final values of the parameters.}
\item{iter}{the number of iterations used.}
\item{conv}{an integer code for the convergence. See \code{convergence} in \code{optim}.}
\item{var}{a variance-covariance matrix of the estimates.}
\item{betavar}{a list of each type of variances of the regression coefficient estimates.}
\item{logalphavar}{a list of each type of variances of the log of baseline parameter estimates.}
\item{thetavar}{a list of each type of variacnes of the regression coefficients and the log of baseline parameter estimates.}
\item{sig2var}{a vector of the variances of random effects variances estimates.}
\item{rhovar}{a vector of the dependence parameter estimates.}
\item{varnames}{a list of event names and variable names.}
\item{control}{a list of control arguments used.}
\item{n}{the number of sample size.}
\item{nevent}{a vector of the number of each types of events.}
\item{neventtype}{the number of types of events.}
\item{nknot}{the number nodes used in Gaussian-quadrature.}
\item{margins}{a vector of character strings specyting marginal distributions of random effects.}
The object will also contain the following: dist, copula, model, call, optionally x, and model.
}
\description{
Fitting a parametric multivariate mixed-Poisson model via a Gaussian copula with a log-normal marginal distribution of random effects
}
